Copilot: move parseSpecTitle to platform/, closing a new creation->catalog edge
This PR's own DI wiring added parseSpecTitle as a value import from
catalog/github-client.ts into creation/creator-code.ts — a new cross-bucket
edge the module-boundary program is trying to eliminate, not add. Copilot
caught it on the current diff.

parseSpecTitle and its parseSpecFrontmatter helper are pure, self-contained
frontmatter parsing with no coupling to catalog's own domain state (catalog's
catalogEntryFromSpec is a consumer, not the owner). Moved both to
platform/spec-frontmatter.ts; github-client.ts now imports and re-exports
parseSpecTitle for its existing same-bucket/root callers, and creator-code.ts
imports it directly from platform, removing the edge entirely.
